pyproject.toml
[project]
name = "codely"
version = "0.1.0"
description = "Lightweight Code Sharing Platform"
authors = [
    {name = "Chukwuka Ibejih"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
    "django-filter (>=24.3,<25.0)",
    "redis (>=5.2.1,<6.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "python-decouple (>=3.8,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
flake8 = "^7.1.1"
black = "^25.1.0"

